+++ app/models/source_editor.rb
@@ -4,4 +4,10 @@ class SourceEditor < Role::SourceRole
     'Editor'
   end
 
+  def year_active_year
+    y = role_object.year
+    y ||= role_object.year_of_publication
+    y
+  end
+
 end